tests for new url functionality in v3 extension
